locomotor/__init__.py: Add function to identify loop and conver to Lua
This function visits each node of the function body. If node is loop and
it performs redis Call then it will convert it into Lua. Lua script will be
stored in ScriptRegistry. NewFunction with loop as body is created which will
take all the required variables as input and returns the variable which are
updated or created in the body.

Signed-off-by: Akash Desai <akashdesai2007@gmail.com>